insights: add background workers which execute search queries and store insights
One challenge here is how to best test this code outside of an actual working dev
server. Some parts (DB actions) are easy to test and I will add tests for them soon,
while others are merely wiring up different components of the system so testing
them is not obvioous.

Stacked on top of #18265

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>